examples/adk/falcon_a2a_agent/requirements.txt
google-adk[a2a]
pyjwt
httpx
python-dotenv
